+++ src/coq_elpi_builtins.ml
@@ -1067,6 +1067,19 @@ It undestands qualified names, e.g. "Nat.t". It's a fatal error if Name cannot b
     state, !: lno +! luno +! (k-1) +? ty, [])),
   DocAbove);
 
+  MLCode(Pred("coq.env.record?",
+    In(inductive, "Ind",
+    Out(bool,"PrimProjs",
+    Read(global, "checks if Ind is a record (PrimProjs = tt if Ind has primitive projections)"))),
+  (fun i _ ~depth {env} _ state ->
+      let mind, indbo = Inductive.lookup_mind_specif env i in
+      match mind.Declarations.mind_record with
+      | Declarations.PrimRecord _ -> !: true
+      | Declarations.FakeRecord -> !: false
+      | _ -> raise No_clause
+    )),
+  DocAbove);
+
   MLCode(Pred("coq.env.const-opaque?",
     In(constant, "GR",
     Read(global, "checks if GR is an opaque constant")),